requirements.txt
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
doclink==1.1.0
fonttools==4.55.3
h11==0.14.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
pmdarima==2.0.4
pycparser==2.22
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
selenium==4.27.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
statsmodels==0.14.4
threadpoolctl==3.5.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
